+++ CHANGES.md
@@ -1,3 +1,9 @@
+Version 1.1.1
+-------------
+
+Unreleased
+
+
 Version 1.1.0
 -------------